#6581: split cycle
