+++ src/utils/inference/selectorInference.ts
@@ -15,14 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {
-  compact,
-  identity,
-  intersection,
-  mapValues,
-  sortBy,
-  uniq,
-} from "lodash";
+import { compact, identity, intersection, sortBy, uniq } from "lodash";
 import { getCssSelector } from "css-selector-generator";
 import { type CssSelectorType } from "css-selector-generator/types/types.js";
 import {
@@ -34,15 +27,12 @@ import { guessUsefulness, isRandomString } from "@/utils/detectRandomString";
 import {
   getSiteSelectorHint,
   SELECTOR_HINTS,
-  type SelectorTemplate,
   type SiteSelectorHint,
 } from "@/utils/inference/siteSelectorHints";
-import { type ElementInfo } from "@/pageScript/frameworks";
-import { type Framework } from "@/pageScript/messenger/constants";
-import * as pageScript from "@/pageScript/messenger/api";
-import { renderString } from "nunjucks";
 import { escapeSingleQuotes, matchesAnyPattern } from "@/utils/stringUtils";
 import { $safeFind } from "@/utils/domUtils";
+import { type ElementInfo } from "@/utils/inference/selectorTypes";
+import { getAttributeSelectorRegex } from "@/utils/inference/selectorInferenceUtils";
 
 export const BUTTON_TAGS: string[] = [
   "li",
@@ -52,6 +42,7 @@ export const BUTTON_TAGS: string[] = [
   "input",
   "svg",
 ];
+
 const MENU_TAGS = ["ul", "tbody"];
 
 export const UNIQUE_ATTRIBUTES: string[] = [
@@ -223,20 +214,6 @@ interface SafeCssSelectorOptions {
   root?: Element;
   excludeRandomClasses?: boolean;
   allowMultiSelection?: boolean;
-}
-
-/**
- * Generates a regex to test attribute selectors generated by `css-selector-generator`,
- * to be used in the whitelist/blacklist arrays. The regex will match any selector with or
- * without a value specified: `[attr]`, `[attr='value']`
- * @example getAttributeSelectorRegex('name', 'aria-label')
- * @returns /^\[name(=|]$)|^\[aria-label(=|]$)/
- */
-export function getAttributeSelectorRegex(...attributes: string[]): RegExp {
-  // eslint-disable-next-line security/detect-non-literal-regexp -- Not user-provided
-  return new RegExp(
-    attributes.map((attribute) => `^\\[${attribute}(=|]$)`).join("|")
-  );
 }
 
 /**
@@ -438,118 +415,6 @@ function findAncestorsWithIdLikeSelectors(
   return $(element).parentsUntil(root).filter(UNIQUE_ATTRIBUTES_SELECTOR).get();
 }
 
-function getElementAttributes(element: HTMLElement) {
-  const attributes: Record<string, string> = {};
-
-  $.each(element.attributes, function () {
-    attributes[this.name] = this.value;
-  });
-
-  return attributes;
-}
-
-function maybeInstantiateSelectorTemplate(
-  ancestorElement: HTMLElement,
-  template: SelectorTemplate
-): string | null {
-  if (!ancestorElement.matches(template.selector)) {
-    return null;
-  }
-
-  const extractedValues = mapValues(template.extract, (extractRule: string) => {
-    const element = $safeFind(extractRule, ancestorElement);
-
-    if (!element) {
-      return null;
-    }
-
-    return {
-      ...getElementAttributes(element[0]),
-      text: element.text(),
-    };
-  });
-
-  // All extraction rules must match, otherwise return no match
-  if (Object.values(extractedValues).some((value) => value == null)) {
-    return null;
-  }
-
-  return renderString(template.template, extractedValues);
-}
-
-/**
- * Return a selector matching one of the provided templates, or null if none of the templates match
- * @param ancestorElement an ancestor of the element to generate the selector for
- * @param templates the selector templates/stencils to try
- * @param root the root element to generate selectors relative to. Used to ensure the generated selector is unique
- */
-function getAncestorSelectorTemplate({
-  ancestorElement,
-  root,
-  templates,
-}: {
-  ancestorElement: HTMLElement;
-  templates: SelectorTemplate[];
-  root?: HTMLElement | Document | JQuery<HTMLElement | Document>;
-}): string | null {
-  // Find first template that matches and uniquely selects the ancestor
-  return templates
-    .map((template) =>
-      maybeInstantiateSelectorTemplate(ancestorElement, template)
-    )
-    .find(
-      (selector) => selector != null && doesSelectOneElement(selector, root)
-    );
-}
-
-function getMatchingRequiredSelectors(
-  element: HTMLElement,
-  requiredSelectors: string[]
-) {
-  return requiredSelectors.find((selector) => element.matches(selector));
-}
-
-/**
- * Return selectors for ancestors of the element based on the matching selector hint, if any hint matches.
- * @param element The element to generate selectors for
- * @param root the root to generate selectors against, either a DOM element or the document
- */
-function mapSelectorOverrideToAncestors(
-  element: HTMLElement,
-  root: HTMLElement | Document
-): Array<{ element: HTMLElement; selectorOverride: string }> {
-  const { requiredSelectors, selectorTemplates } = getSiteSelectorHint(element);
-
-  const ancestors =
-    root === document
-      ? $(element).parents().get()
-      : $(element)
-          .parentsUntil(root as HTMLElement)
-          .get();
-
-  // Ancestors from root to element
-  ancestors.reverse();
-
-  let currentRoot = root;
-  const result = [];
-
-  for (const ancestor of ancestors) {
-    const selectorOverride =
-      getAncestorSelectorTemplate({
-        ancestorElement: ancestor,
-        root: currentRoot,
-        templates: selectorTemplates,
-      }) ?? getMatchingRequiredSelectors(ancestor, requiredSelectors);
-
-    if (selectorOverride) {
-      currentRoot = ancestor;
-      result.push({ element: ancestor, selectorOverride });
-    }
-  }
-
-  return result;
-}
-
 export function inferSelectorsIncludingStableAncestors(
   element: HTMLElement,
   root?: Element,
@@ -619,7 +484,7 @@ export function inferSelectors(
   );
 }
 
-type InferSelectorArgs = {
+export type InferSelectorArgs = {
   /**
    * The root to generate the selector with respect to. Provide null to use the document.
    */
@@ -629,107 +494,6 @@ type InferSelectorArgs = {
    */
   excludeRandomClasses: boolean;
 };
-
-/**
- * @deprecated framework-based (e.g., React/Vue/etc.) traversal is deprecated
- */
-type FrameworkSelectorArgs = {
-  /**
-   * The front-end framework to assume for component traversal.
-   */
-  framework?: Framework;
-  /**
-   * Number of non-DOM components to traverse up.
-   */
-  traverseUp: number;
-};
-
-/**
- * Infers valids selectors for a single element.
- * @see inferMultiElementSelector
- */
-export async function inferSingleElementSelector({
-  element,
-  root,
-  excludeRandomClasses,
-  framework,
-  traverseUp,
-}: InferSelectorArgs &
-  FrameworkSelectorArgs & {
-    element: HTMLElement;
-  }): Promise<ElementInfo> {
-  if (framework) {
-    // We're using pageScript getElementInfo only when specific framework is used.
-    // On Salesforce we were running into an issue where certain selectors weren't finding any elements when
-    // run from the pageScript. It might have something to do with the custom web components Salesforce uses?
-    // In any case, the pageScript is not necessary if framework is not specified, because selectElement
-    // only needs to return the selector alternatives.
-    const selector = safeCssSelector([element], {
-      excludeRandomClasses,
-    });
-
-    // XXX: pageScript.getElementInfo doesn't support `root` because we can't pass an element across the pageScript
-    // boundary.
-    return pageScript.getElementInfo({
-      selector,
-      framework,
-      traverseUp,
-    });
-  }
-
-  // Ancestors in order from root to element
-  const ancestorSelectorOverrides = mapSelectorOverrideToAncestors(
-    element,
-    root
-  );
-
-  const ancestorSelectors = ancestorSelectorOverrides.map(
-    ({ selectorOverride }) => selectorOverride
-  );
-
-  const rootOverride =
-    // Scope the element generation to the innermost ancestor with a selector override
-    ancestorSelectorOverrides.length > 0
-      ? ancestorSelectorOverrides.at(-1).element
-      : root;
-
-  const selectorWithRootOverride = safeCssSelector([element], {
-    root: rootOverride,
-    excludeRandomClasses,
-  });
-
-  // Prepend all the selectors with the required selectors
-  const inferredSelectors = uniq(
-    [
-      selectorWithRootOverride,
-      ...inferSelectorsIncludingStableAncestors(element, rootOverride),
-    ].map((selector) => [...ancestorSelectors, selector].join(" "))
-  );
-
-  // Filter out any malformed selectors and/or selectors that don't exactly match the element
-  const validatedSelectors = inferredSelectors.filter((selector) => {
-    try {
-      const match = $safeFind(selector, root);
-      return match.length === 1 && match.get(0) === element;
-    } catch {
-      console.warn("Invalid selector", selector);
-      // Catch invalid selectors
-      return false;
-    }
-  });
-
-  if (validatedSelectors.length === 0) {
-    throw new Error("Automatic selector generation failed");
-  }
-
-  return {
-    selectors: validatedSelectors,
-    framework: null,
-    hasData: false,
-    tagName: element.tagName,
-    parent: null,
-  };
-}
 
 /**
  * Infer a selector that matches multiple elements.
@@ -777,7 +541,7 @@ export function inferMultiElementSelector({
 /**
  * Returns true if selector uniquely identifies a single element on the page.
  */
-function doesSelectOneElement(
+export function doesSelectOneElement(
   selector: string,
   parent?: HTMLElement | Document | JQuery<HTMLElement | Document>
 ): boolean {
